fix ghost location persistence bug
when two cards 1 & 2 are stacked on top of each other, they
are stored in a deck called deck-1.

If this deck is dragged to the player's hand (say), then
the ghost of deck-1 is recorded as being over the hand.

If the deck is subsequently recreated (by dragging card 1
to the table followed by card 2 over it), then crossfade.ts
things that deck-1 moved from the hand to the table (because
the ephemeral decks are created using stable ID's).

This commit fixes the problem by expiring the ghost's location
after 500ms (enough for any animations to run that actually
need it).